Refactor to allow more easily reporting that Java is not installed
If the appropriate Win32Exception is thrown, a new JavaNotFoundException is thrown. This is handled by the caller and reported through the newly-exposed method to report errors in projects. This is then passed to the errors window.

This also moves process management away from SkyEditor.ROMEditor.Windows, a project intended to only house things that are Windows-specific, and Java is cross-platform and platform-agnostic.